move playbooks to a separate directory move test executables to `bin`
