+++ scripts/concent_integration_tests/bin/additional_verification.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
-from scripts.concent_integration_tests.tests.playbooks import (
-    additional_verification, run
+from scripts.concent_integration_tests.playbooks import (
+    run, additional_verification
 )
 
 run.run_playbook(additional_verification.AdditionalVerification)  # type: ignore
