+++ scripts/concent_integration_tests/bin/force_accept.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-from scripts.concent_integration_tests.tests.playbooks import force_accept, run
-
+from scripts.concent_integration_tests.playbooks import run, force_accept
 
 run.run_playbook(force_accept.ForceAccept)  # type: ignore
