+++ scripts/concent_integration_tests/bin/force_report.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
-from scripts.concent_integration_tests.tests.playbooks import force_report, run
-
+from scripts.concent_integration_tests.playbooks import run, force_report
 
 run.run_playbook(force_report.ForceReport)  # type: ignore
