+++ scripts/concent_integration_tests/bin/no_concent.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+from scripts.concent_integration_tests.playbooks import run, no_concent
+
+run.run_playbook(no_concent.NoConcent)  # type: ignore
