+++ scripts/concent_integration_tests/bin/regular_task_run.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-from scripts.concent_integration_tests.tests.playbooks import regular_run, run
-
+from scripts.concent_integration_tests.playbooks import run, regular_run
 
 run.run_playbook(regular_run.RegularRun)  # type: ignore
