+++ scripts/concent_integration_tests/bin/rpc.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+from scripts.concent_integration_tests.playbooks import run, rpc_test
+
+run.run_playbook(rpc_test.RPCTest)  # type: ignore
