+++ scripts/concent_integration_tests/bin/rpc_mainnet.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-from scripts.concent_integration_tests.tests.playbooks import rpc_test, run
+from scripts.concent_integration_tests.playbooks import run, rpc_test
 from golem.config.environments import set_environment
 
 set_environment('mainnet', 'disabled')
