+++ scripts/concent_integration_tests/bin/rpc_no_concent.py
@@ -1,4 +1,4 @@
 #!/usr/bin/env python
-from scripts.concent_integration_tests.tests.playbooks import rpc_test, run
+from scripts.concent_integration_tests.playbooks import run, rpc_test
 
 run.run_playbook(rpc_test.NoConcentRPCTest)  # type: ignore
