+++ scripts/concent_integration_tests/bin/timeout.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-from scripts.concent_integration_tests.tests.playbooks import task_timeout, run
-
+from scripts.concent_integration_tests.playbooks import run, task_timeout
 
 run.run_playbook(task_timeout.TaskTimeoutAndRestart)  # type: ignore
