+++ scripts/concent_integration_tests/playbooks/additional_verification.py
@@ -1,7 +1,5 @@
 from scripts.concent_integration_tests import helpers
-from scripts.concent_integration_tests.tests.playbooks.base import (
-    NodeTestPlaybook
-)
+from .base import NodeTestPlaybook
 
 
 class AdditionalVerification(NodeTestPlaybook):
