+++ scripts/concent_integration_tests/playbooks/force_report.py
@@ -1,14 +1,10 @@
 from scripts.concent_integration_tests import helpers
-from scripts.concent_integration_tests.tests.playbooks.base import NodeTestPlaybook
+from .base import NodeTestPlaybook
 
 
-class ForceDownload(NodeTestPlaybook):
-    provider_node_script = 'provider/debug'
-    requestor_node_script = 'requestor/fail_results'
-
-    def step_clear_requestor_output(self):
-        helpers.clear_output(self.requestor_output_queue)
-        self.next()
+class ForceReport(NodeTestPlaybook):
+    provider_node_script = 'provider/impatient_frct'
+    requestor_node_script = 'requestor/no_ack_rct'
 
     def step_clear_provider_output(self):
         helpers.clear_output(self.provider_output_queue)
@@ -16,18 +12,8 @@ class ForceDownload(NodeTestPlaybook):
 
     def step_wait_task_finished(self):
         concent_fail = helpers.search_output(
-            self.requestor_output_queue,
-            '.*Concent request failed.*',
-        )
-
-        if concent_fail:
-            print("Requestor: ", concent_fail.group(0))
-            self.fail()
-            return
-
-        concent_fail = helpers.search_output(
             self.provider_output_queue,
-            ".*Concent request failed.*|.*Can't receive message from Concent.*",
+            '.*Concent request failed.*|.*Problem interpreting.*',
         )
 
         if concent_fail:
@@ -46,7 +32,6 @@ class ForceDownload(NodeTestPlaybook):
         NodeTestPlaybook.step_wait_provider_gnt,
         NodeTestPlaybook.step_wait_requestor_gnt,
         NodeTestPlaybook.step_get_known_tasks,
-        step_clear_requestor_output,
         step_clear_provider_output,
         NodeTestPlaybook.step_create_task,
         NodeTestPlaybook.step_get_task_id,
