+++ scripts/concent_integration_tests/playbooks/no_concent.py
@@ -1,4 +1,4 @@
-from scripts.concent_integration_tests.tests.playbooks.base import NodeTestPlaybook
+from .base import NodeTestPlaybook
 
 
 class NoConcent(NodeTestPlaybook):
