+++ scripts/concent_integration_tests/playbooks/regular_run.py
@@ -1,9 +1,6 @@
-from scripts.concent_integration_tests.tests.playbooks.base import (
-    NodeTestPlaybook
-)
+from .base import NodeTestPlaybook
 
 
 class RegularRun(NodeTestPlaybook):
     provider_node_script = 'provider/debug'
     requestor_node_script = 'requestor/debug'
-    task_settings = 'jpeg'
