--- scripts/concent_integration_tests/tests/test_force_download.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python
-import sys
-from scripts.concent_integration_tests.tests.playbooks import (
-    force_download, run
-)
-
-
-run.run_playbook(force_download.ForceDownload)  # type: ignore
